feat(transform): make transformations use integers instead of floats (#451)
* feat(transform): make transformations use integers instead of floats

Transformations now only encode Manhattan transformations
and internally store a transformation matrix with integer (i8)
entries, as opposed to f64 entries.

* remove approx from cargo toml

+++ libs/atoll/src/abs.rs
@@ -657,7 +657,7 @@ impl<PDK: Pdk> Draw<PDK> for &DebugAbstract {
                             let text = Text::new(
                                 layer_id,
                                 format!("({x},{y})"),
-                                Transformation::translate(pt.x as f64, pt.y as f64),
+                                Transformation::translate(pt.x, pt.y),
                             );
                             recv.draw(text)?;
                         }
